Change temperature type from f64 to i32

+++ src/applet.rs
@@ -17,7 +17,7 @@ struct Weather {
     popup: Option<cosmic::iced::window::Id>,
     config: WeatherConfig,
     config_handler: Option<cosmic::cosmic_config::Config>,
-    temperature: f64,
+    temperature: i32,
     latitude: String,
     longitude: String,
     use_fahrenheit: bool,
@@ -36,7 +36,7 @@ impl Weather {
                 }
                 Err(error) => {
                     tracing::error!("Failed to get location forecast: {error:?}");
-                    cosmic::action::Action::App(Message::UpdateTemperature(0.0))
+                    cosmic::action::Action::App(Message::UpdateTemperature(0))
                 }
             },
         )
@@ -44,7 +44,7 @@ impl Weather {
 
     fn format_temperature(&self) -> String {
         if self.use_fahrenheit {
-            format!("{:.1}°F", self.temperature * 9.0 / 5.0 + 32.0)
+            format!("{:.1}°F", self.temperature * 9 / 5 + 32)
         } else {
             format!("{:.1}°C", self.temperature)
         }
@@ -56,7 +56,7 @@ pub enum Message {
     Tick,
     ToggleWindow,
     PopupClosed(cosmic::iced::window::Id),
-    UpdateTemperature(f64),
+    UpdateTemperature(i32),
     UpdateLatitude(String),
     UpdateLongitude(String),
     ToggleFahrenheit(bool),
@@ -83,7 +83,7 @@ impl cosmic::Application for Weather {
                 popup: None,
                 config: flags.config,
                 config_handler: flags.config_handler,
-                temperature: 0.0,
+                temperature: 0,
                 latitude: latitude.to_string(),
                 longitude: longitude.to_string(),
                 use_fahrenheit,
