+++ src/weather.rs
@@ -92,7 +92,7 @@ struct Next12HoursDetails {
 pub async fn get_location_forecast(
     latitude: String,
     longitude: String,
-) -> Result<f64, reqwest::Error> {
+) -> Result<i32, reqwest::Error> {
     let url = format!(
         "https://api.met.no/weatherapi/locationforecast/2.0/compact?lat={latitude}&lon={longitude}",
     );
@@ -109,8 +109,8 @@ pub async fn get_location_forecast(
         .properties
         .timeseries
         .first()
-        .map(|d| d.data.instant.details.air_temperature)
-        .unwrap_or(0.0);
+        .map(|d| d.data.instant.details.air_temperature as i32)
+        .unwrap_or(0);
 
     Ok(current_temperature)
 }
